data: Move batch file scope resolution
Batch selection helpers validate line IDs and translate display IDs for
batch-source operations. File-scope resolution also exists there, and it
narrows pathless operations through selected-change session state.

The project does not provide a session-owned place for batch file-scope
resolution. Keeping that path in the batch package makes batch selection
depend on selected batch binaries, selected batch gitlinks, and selected
change paths.

This commit addresses that by moving batch file-scope resolution into
data.batch_file_scope and updating command and file-review callers to use
that session-aware helper. batch.selection keeps line-selection and
ownership-selection primitives.

The final commit will encode that boundary in the architecture tests.
