+++ src/git_stage_batch/commands/batch_source/reset_claims.py
@@ -18,7 +18,7 @@ from ...batch.ownership_unit_rebuild import rebuild_ownership_from_units
 from ...batch.ownership_unit_selection import filter_ownership_units_by_display_ids
 from ...batch.ownership_unit_validation import validate_ownership_units
 from ...batch.query import read_batch_metadata
-from ...batch.selection import require_display_ids_available, resolve_batch_file_scope
+from ...batch.selection import require_display_ids_available
 from ...batch.state_refs import sync_batch_state_refs
 from ...batch.text_file_storage import (
     add_file_to_batch,
@@ -33,6 +33,7 @@ from ...batch.submodule_pointer import (
 )
 from ...batch.validation import batch_exists
 from ...core.line_selection import LineRanges
+from ...data.batch_file_scope import resolve_batch_file_scope
 from ...utils.repository_buffers import load_git_object_as_buffer
 from ...exceptions import MergeError, exit_with_error
 from ...i18n import _
