+++ src/git_stage_batch/commands/show_from.py
@@ -11,12 +11,10 @@ from .batch_source import file_list_action as _file_list_action
 from .batch_source import replacement_previews as _replacement_previews
 from ..batch.metadata_validation import read_validated_batch_metadata
 from ..core.replacement import ReplacementPayload
-from ..batch.selection import (
-    resolve_batch_file_scope,
-    require_single_file_context_for_line_selection,
-)
+from ..batch.selection import require_single_file_context_for_line_selection
 from ..batch.source_selector import parse_batch_source_selector
 from ..batch.validation import batch_exists
+from ..data.batch_file_scope import resolve_batch_file_scope
 from ..data.file_review.batch_selection import translate_batch_file_gutter_ids_to_selection_ids
 from ..exceptions import (
     exit_with_error,
