+++ src/git_stage_batch/data/file_review/batch_selection.py
@@ -7,10 +7,7 @@ from dataclasses import replace
 from typing import TYPE_CHECKING
 
 from ...batch.file_display import render_batch_file_display
-from ...batch.selection import (
-    require_single_file_context_for_line_selection_ranges,
-    resolve_batch_file_scope,
-)
+from ...batch.selection import require_single_file_context_for_line_selection_ranges
 from ...batch.submodule_pointer import (
     is_batch_submodule_pointer,
     refuse_batch_submodule_pointer_lines,
@@ -18,6 +15,7 @@ from ...batch.submodule_pointer import (
 from ...core.line_selection import LineRanges
 from ...exceptions import exit_with_error
 from ...i18n import _
+from ..batch_file_scope import resolve_batch_file_scope
 from .batch_selection_freshness import fresh_batch_review_selections_for_action
 from .records import FileReviewAction
 from .selection_validation import validate_review_scoped_line_selection
